[WIP] Support SavedStateHandle.getLiveData / getStateFlow using keys (#1)
This kinda sucks because before this a `Key` was kind of simple and
elegant with only get & set methods. However if we want to support the
stateFlow and liveData methods in SavedStateHandle, we need to split up
& capture the backing data type (this should also be useful for
supporting the compose nav case down the road).

Steps:
 - Add `BACKED_BY: Any?` type to Key and AsyncKey
- Split keys' get & set into getBackingData & mapGet + setBackingData &
mapSet
- add generic method to create a `NativeKey` which is just any Key where
types `T === BACKED_BY`
- add support for new methods on saveStateHandle leveraging mapGet &
mapSet

Piggyback: As an optimization, we make Key a data class (we can haz
generic data classes now, yay). If we revert this PR we should keep this
change.