Start work on nav compose module (#2)
This PR starts work on a nav compose module and the refactors needed to
make it work. We do not finish the module here, however we did figure
out the syntax & reflection complexities, so for now (since this whole
repo is still in a prototype phase) we're just gonna merge this bad boy
and keep moving.